multiagent/requirements.txt
Flask
requests
beautifulsoup4
tavily-python